change return type of next function in FP unit

+++ structures/functional_units.h
@@ -27,7 +27,7 @@ public:
 	int right_operand;
 	Instruction* instr;
 	void dispatch(Instruction* instr);
-	int next();
+	double next();
 	bool occupied;
 	FPFunctionalUnit(int cycles_ex);
 };
